Recaptcha: -updated downloadCaptcha to return null because download will fail with 400 Bad Request ChallengeResponseController: -since RC1 shutdown only a dummy response with two words is required PluginForHost/PluginForDecrypt: -updated copyCaptcha to handle null Plugin: -updated clean to handle null TwentyOneMembersCom: -fixed login -fixes ticket GGE-671-15676 build>full
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@39062 ebf7c1c2-ba36-0410-9fe8-c592906822b4

Former-commit-id: e75a538b3eb2ea53f8f7d200624f96ae2827e757

+++ src/jd/plugins/PluginForDecrypt.java
@@ -564,11 +564,15 @@ public abstract class PluginForDecrypt extends Plugin {
     }
 
     private File copyCaptcha(String method, File file) throws Exception {
-        final File copy = Application.getResource("captchas/" + method + "/" + Hash.getMD5(file) + "." + Files.getExtension(file.getName()));
-        copy.delete();
-        copy.getParentFile().mkdirs();
-        IO.copyFile(file, copy);
-        return copy;
+        if (file != null) {
+            final File copy = Application.getResource("captchas/" + method + "/" + Hash.getMD5(file) + "." + Files.getExtension(file.getName()));
+            copy.delete();
+            copy.getParentFile().mkdirs();
+            IO.copyFile(file, copy);
+            return copy;
+        } else {
+            return null;
+        }
     }
 
     protected ClickedPoint getCaptchaClickedPoint(String method, File file, final CryptedLink link, String defaultValue, String explain) throws Exception {
